// src/app/api/models/medication.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Detalles del Medication a crear.
 */
export interface Medication {

  /**
   * Status of the medication
   */
  active?: boolean;

  /**
   * Unique code of the medication
   */
  code: string;

  /**
   * Commercial Name of the medication
   */
  commercialName?: string;

  /**
   * Concentration level
   */
  concentration?: string;

  /**
   * Generic Name of the medication
   */
  genericName: string;

  /**
   * Identificador único
   */
  id?: number;

  /**
   * Pharmaceutical Form
   */
  pharmaceuticalForm?: string;

  /**
   * Presentation
   */
  presentation?: string;
}